picchat LiveView authorization

+++ example_projects/pic_chat/lib/pic_chat_web/live/message_live/index.ex
@@ -5,7 +5,7 @@ defmodule PicChatWeb.MessageLive.Index do
   alias PicChat.Chat.Message
 
   @impl true
-  def mount(_params, _session, socket) do
+  def mount(_params, session, socket) do
     {:ok, stream(socket, :messages, Chat.list_messages())}
   end
 
@@ -40,8 +40,12 @@ defmodule PicChatWeb.MessageLive.Index do
   @impl true
   def handle_event("delete", %{"id" => id}, socket) do
     message = Chat.get_message!(id)
-    {:ok, _} = Chat.delete_message(message)
 
-    {:noreply, stream_delete(socket, :messages, message)}
+    if message.user_id == socket.assigns.current_user.id do
+      {:ok, _} = Chat.delete_message(message)
+      {:noreply, stream_delete(socket, :messages, message)}
+    else
+      {:noreply, Phoenix.LiveView.put_flash(socket, :error, "You are not authorized to delete this message.")}
+    end
   end
 end
